Handle (partially) .nomedia cases, fixes #29
Fixes cases when image is modified inside some directory which has
.nomedia in it, or some of its ancestors have .nomedia.

Also fixes case when .nomedia file itself is deleted, as this means all
images from this directory (and child dirs) are now eligable for face
detection. However, similar case (when .nomedia is added to some
directory) is not handled and is just referenced as issue #37. This one
requires new task to be added.