Central place for loading DNN models, fixes #15
This change adds support to download models in Makefile (and adds it as
default target in make).

Requirement task is renamed to check not only does pdlib is loaded, but
also if model files are present.

Requirement class is changed, so modelId is required argument for it to
operate. It encapsulate logic to retrieve paths to model files based on
model ID (which currently does not make any sense:) as we have only one
model), but yes - Requirement now hides this from developer.