Add exceptions to background tasks
There was couple of todos for this. Basically, idea was that task throw
exceptions on errors and that main event loop for tasks handle that.
Some tasks are already throwing exceptions, some are throwing now,
but yes - control flow should be with exceptions, not with return values.

I also tried to fix scrutinizer errors from files that I already modified.