Fix for non-existing image
If image doesn't exist on disk, but do exist in Nextcloud filecache,
we don't see it, and we don't want to add it to our database
(by-design).
So, we signal that processing context is null, but code had bug
(regression actually) where it did not handle null case. This fixes it.