Set person as valid after cluster recreation
When we were doing cluster recreation, only new clusters were getting
`is_valid` set as true. Existing (but changed) clusters was not getting
updated value for `is_valid`, so if it was false, it remained false.
This triggered new cluster recreations (as logic was thinking there are
stale persons and cluster recreation is needed) in infinite loop.

This change adds update to DB, so that each modified cluster is set as
valid.