Deletes persons without faces on change, fixes #137
Using approach #1 from #137.

This change removes all persons that are without faces, after image
(file) is modified/deleted in hooks. When file is modified/deleted, we
are invalidating all persons for which there are faces on that image and
then removing faces. If this is last face for a given person, we didn't
delete that person, but we should (as there is no point having it), so
this fixed that.