Remove Nextcloud 15 support
It is not a happy decision, in our tests it works from NC14 to NC17 without
changes.

The requirement of the Nexcloud App Store to not use deprecated data types
requires us to change a simple column type from 'json_array', to 'json' but
this also makes it incompatible with nexcloud less than 16.

Although it would be complicated we could make it compatible, but since
Nextcloud 15 is close to its end of life, we prefer to cut for the healthy
and not support NC15.

If you want to reverse it, we can help you, but we suggest upgrading to NC 16.
:smiley: