Add a custom way to disable analysis in folders. Issue #171
This is done as described in issue 'Per-directory settings support #89'
But it's just a small part of it.

Then you have to add a .facerecognition.json file with '{detection = off}' in
the folder you want to ignore. It's complicated, but then we'll add a side
panel to do it easily.

Note that the .nomedia file will still be considered. If there is a nomedia file,
these folders will be ignored directly from the analysis.