Suggest all proposals, even if they already have the same name.
Then filter (Auto accepting or rejecting the proposals) while you accept the
proposals.

This ensures that you can apply as many relations as possible, but must accept
fewer suggestions.

On the other hand, now not apply any change until finish with all the
suggestions. So, if you close the dialog no changes will be applied.