First force, then security controls, then true conditions, and finally the default response.
Avoid denying so many conditions that it becomes unnecessarily more confusing.

Change condition for the initial creation of the clusters to:
* Ideally, 95% of the photos analyzed.
* or at least 1000 faces found.

For the recreation of the clusters, the minimum number of faces is increased
from 10 to 25. But always after 2 hours from the oldest face also recreates it.

Note that this is independent of the amount of user photos. But perhaps we
should think in percentages of the total. Something to consider.