Introduce a new model v4: DLibCnnHogModel.
Ok.  Basically it's a spawn. I recognize it. hahah

Well, according to the last experiments, the weak link in our default model is
the landmark detector that can't align profile faces correctly to compare them
accurately.

The Hog detector also fails and cannot detect these faces.
So, if Hog detects it, we can assume that the landmark detector will do it too.
and this would be a good face..

If Hog don't detect this face, they are probably in profile or rotated.
These are bad to compare, so just we lower the confidence a little, to avoid
clustering them.