Show only the most relevant information on admin page.
Not show the version of the pdlib or if it is not installed. If it is not
installed, you will not be able to install the application.

Neither does the system memory. Only matther the memory assigned, or if you
haven't done it yet.

The checks that I just remove here are controlled by the two remaining.