Add a command to search for duplicate images.
Everyone knows that my test suite is a collection of images from The
Bing Bang Theory. I always download more images to make it more real,
but it is impossible to know if the images are duplicated or not.. :)

The logic here is simple, if two images are the same, they have the
same faces, then their comparison should be identical.