Add support for using imaginary to create the temporary files.
This allows for "paratically" free add support for images heic,
tiff, and many more.

You must configure the imaginary service in nextcloud.

See the documentation about it, but basically you need to edit
config/config.php and append:

+'preview_imaginary_url' => 'http://localhost:9000',

...and enable new file types in our app:

+'enabledFaceRecognitionMimetype' => array(
+  0 => 'image/jpeg',
+  1 => 'image/png',
+  2 => 'image/heic',
+  3 => 'image/tiff',
+  4 => 'image/webp',
+),

Finally adds the --crawl-missing option that forces a search for the
files to analyze the new allowed types.