Add other execution modes to the background_job command.
Without special arguments, it is executed as before, designed for
progressive work.

With the --sync-mode argument, only the file synchronization tasks are
executed, and this is done forcefully. We could say that it replaces
the --crawl-missing option.
With the --analyze-mode argument it only executes the task of
analyzing the images to find the faces.
Finally, with the --cluster-mode option, only the task of face
clustering is executed.

The --defer-clustering argument still works the same way.