Makes --analyze-mode thread-safe. Issue #237
At least it's trying. Basically this mode implements a file-level lock
and can be executed in parallel without working twice on the same
file.

To do this, the global lock is ignored. You can perform a reset while
doing multiple executions of this command. Maybe I should add an
option to make it more explicit.