Add the sixth model to the application. Aka DlibTaguchiHog model. =)
In this model 6, we use the model newly trained by Tanguchi to obtain
the descriptor and perform the face recognition.

This model was trained from scratch to slightly improve the bias of the
original model on non-Caucasian/American people. It obtained a similar
result in the LFW tests, slightly lower, but within the acceptable
margins of error.

It may still have bias problems, they recognize that it fails with
African people and with childrens, but in principle it seems like an
improvement to the model used in 1, 2, 3 and 4. :wink:

See: https://github.com/TaguchiModels/dlibModels

This model is added for testing, and I understand that it will be of
great help to those who have had problems with the original model.