Improves the review ignored people button introduced in #747
It is a minimal improvement on how the application should always have
been. Instead of getting the information at the start of the
application, we do it only when we know there are values and the user
wants to do something with it.

This also applies to the user's enablement value as well as to the set
of new people to recognize.