Adds batch option for face clustering. Issue #712 (Among many others)
This improves the performance of clustering with many faces. In my
tests with 43 thousand faces, it takes 11,91 minutes to run the full
'face:background_job --cluster-mode' command. Using a batch size of 20
thousand the time is reduced to 4.47 minutes. With 5000, 1.71 minute
and with 2000 (Which is the minimum cut it takes only 54 seconds.

Against all odds, memory consumption does not increase in any way,
but it still has a disadvantage. The clusters... although they do not
increase their quantity as much (only 5%), these will generally be
smaller (Presumably of higher quality), but there will be more
clusters to give them names.

Well, this is another advanced option that will not be available in
the administrator interface.

occ config:app:set  facerecognition clustering_batch_size --value='1000' --type=integer